Added Sine functions, fixed cut functions
Added support for sine functions using the ~ symbol.
Before when the function had a extreme gradient it was cut in the window because only points in the defined area were rendered. Now the points just arent excluded because the java frame supports using coordinates outside the window when drawing lines and points.